First Attempt
A little bit of nose-diving into the MVC code.
Added the ‘ENTER CREATIVE CASTLE’ button on the nav (both
pre-sticky-menu and post-sticky-menu).

Noticed that the button disappears after the story, so it may be script
related.

Lastly, I changed the SASS variable for $contentWidth to 1200px to make
the new button fit.